requirements.txt
azure-storage-blob==12.20.0
coloredlogs==15.0.1
matplotlib==3.9.2
pandas==2.2.3
python-dotenv==1.0.1
rioxarray==0.16.0
tqdm==4.66.4
xarray==2024.3.0
dask==2024.7.0
scipy==1.14.1
pre-commit==3.8.0
netCDF4==1.7.2
